Added BaseAuto that other classes extend from
