+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/OpenCVDetection.java
@@ -61,7 +61,6 @@ public abstract class OpenCVDetection extends LinearOpMode {
 
         while (opModeIsActive())
         {
-            telemetry.addData("Analysis", pipeline.getAnalysis());
             telemetry.addData("", pipeline.position);
             telemetry.update();
